requirements.txt
aiomysql==0.1.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
Authlib==1.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
dynaconf==3.2.4
ecdsa==0.18.0
fastapi==0.95.2
fastapi-authlib==0.0.4
fastapi-sa==0.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.21.1
